__local_op bugfix

+++ heat/core/tests/test_indexing.py
@@ -12,7 +12,7 @@ class TestIndexing(unittest.TestCase):
         cond = a > 3
         nz = ht.nonzero(cond)
         self.assertEqual(nz.gshape, (5, 2))
-        self.assertEqual(nz.dtype, ht.int)
+        self.assertEqual(nz.dtype, ht.int64)
         self.assertEqual(nz.split, None)
 
         # split
@@ -20,7 +20,7 @@ class TestIndexing(unittest.TestCase):
         cond = a > 3
         nz = cond.nonzero()
         self.assertEqual(nz.gshape, (6, 2))
-        self.assertEqual(nz.dtype, ht.int)
+        self.assertEqual(nz.dtype, ht.int64)
         self.assertEqual(nz.split, 0)
         a[nz] = 10.
         self.assertEqual(ht.all(a[nz] == 10), 1)
@@ -32,14 +32,14 @@ class TestIndexing(unittest.TestCase):
         cond = a > 3
         wh = ht.where(cond)
         self.assertEqual(wh.gshape, (6, 2))
-        self.assertEqual(wh.dtype, ht.int)
+        self.assertEqual(wh.dtype, ht.int64)
         self.assertEqual(wh.split, None)
         # split
         a = ht.array([[1, 2, 3], [4, 5, 6], [7, 8, 9]], split=1)
         cond = a > 3
         wh = ht.where(cond)
         self.assertEqual(wh.gshape, (6, 2))
-        self.assertEqual(wh.dtype, ht.int)
+        self.assertEqual(wh.dtype, ht.int64)
         self.assertEqual(wh.split, 0)
 
         # not split cond
